+++ internal/app/jsonStructs.go
@@ -1,0 +1,34 @@
+package app
+
+import "encoding/json"
+
+type JsonPayload interface {
+	Marshal() (result []byte, err error)
+}
+
+type RequestJson struct {
+	Expression string `json:"expression"`
+}
+
+func (r RequestJson) Marshal() (result []byte, err error) {
+	result, err = json.Marshal(&r)
+	return
+}
+
+type OKJson struct {
+	Result float64 `json:"result"`
+}
+
+func (o OKJson) Marshal() (result []byte, err error) {
+	result, err = json.Marshal(&o)
+	return
+}
+
+type ErrorJson struct {
+	Error string `json:"error"`
+}
+
+func (e ErrorJson) Marshal() (result []byte, err error) {
+	result, err = json.Marshal(&e)
+	return
+}
