+++ internal/app/server.go
@@ -1,0 +1,13 @@
+package app
+
+import (
+	"github.com/Debianov/calc-ya-go-24/internal"
+	"net/http"
+)
+
+func StartServer() (err error) {
+	http.HandleFunc("/api/v1/calculate", CalcHandler)
+	s := internal.GetDefaultServer()
+	err = s.ListenAndServe()
+	return
+}
