+++ internal/config.go
@@ -1,1 +1,7 @@
 package internal
+
+import "net/http"
+
+func GetDefaultServer() *http.Server {
+	return &http.Server{Addr: "127.0.0.1:8000"}
+}
